chore(core-manager): move logic form core-watcher to core-manager
