--- packages/core-watcher/.gitattributes
@@ -1,7 +1,0 @@
-# Path-based git attributes
-# https://www.kernel.org/pub/software/scm/git/docs/gitattributes.html
-
-# Ignore all test and documentation with "export-ignore".
-/.gitattributes   export-ignore
-/.gitignore       export-ignore
-/README.md        export-ignore
