--- packages/core-watcher/README.md
@@ -1,21 +1,0 @@
-# ARK Core - Watcher
-
-<p align="center">
-    <img src="https://raw.githubusercontent.com/ARKEcosystem/core/master/banner.png" />
-</p>
-
-## Documentation
-
-You can find installation instructions and detailed instructions on how to use this package at the [dedicated documentation site](https://ark.io/documentation).
-
-## Security
-
-If you discover a security vulnerability within this package, please send an e-mail to security@ark.io. All security vulnerabilities will be promptly addressed.
-
-## Credits
-
-This project exists thanks to all the people who [contribute](../../../../contributors).
-
-## License
-
-[MIT](LICENSE) © [ARK Ecosystem](https://ark.io)
