--- packages/core-watcher/package.json
@@ -1,41 +1,0 @@
-{
-    "name": "@arkecosystem/core-watcher",
-    "version": "3.0.0-next.0",
-    "description": "Manager for ARK Core",
-    "license": "MIT",
-    "contributors": [
-        "Sebastijan Kuzner <sebastijan@ark.io>"
-    ],
-    "files": [
-        "dist"
-    ],
-    "main": "dist/index",
-    "types": "dist/index",
-    "scripts": {
-        "build": "yarn clean && yarn compile",
-        "build:watch": "yarn clean && yarn compile -w",
-        "build:docs": "../../node_modules/typedoc/bin/typedoc --out docs src",
-        "clean": "del dist",
-        "compile": "../../node_modules/typescript/bin/tsc",
-        "prepublishOnly": "yarn build",
-        "pretest": "bash ../../scripts/pre-test.sh"
-    },
-    "dependencies": {
-        "@arkecosystem/core-kernel": "^3.0.0-next.0",
-        "better-sqlite3": "^7.0.0"
-    },
-    "devDependencies": {
-        "nock": "^12.0.3"
-    },
-    "engines": {
-        "node": ">=10.x"
-    },
-    "publishConfig": {
-        "access": "public"
-    },
-    "arkecosystem": {
-        "core": {
-            "alias": "watcher"
-        }
-    }
-}
