--- packages/core-watcher/src/defaults.ts
@@ -1,3 +1,0 @@
-export const defaults = {
-    storage: `${process.env.CORE_PATH_DATA}/events.sqlite`,
-};
