--- packages/core-watcher/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./database-service";
-export * from "./service-provider";
