+++ packages/core/bin/config/testnet/app.json
@@ -38,7 +38,10 @@
                 "package": "@arkecosystem/core-forger"
             },
             {
-                "package": "@arkecosystem/core-watcher"
+                "package": "@arkecosystem/core-snapshots"
+            },
+            {
+                "package": "@arkecosystem/core-manager"
             }
         ]
     },
@@ -106,9 +109,6 @@
             },
             {
                 "package": "@arkecosystem/core-snapshots"
-            },
-            {
-                "package": "@arkecosystem/core-watcher"
             },
             {
                 "package": "@arkecosystem/core-manager"
